# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrivacyRequestCreate": {
        "properties": {
          "details": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "request_type": {
            "enum": [
              "access",
              "correction",
              "deletion"
            ],
            "title": "Request Type",
            "type": "string"
          }
        },
        "required": [
          "request_type"
        ],
        "title": "PrivacyRequestCreate",
        "type": "object"
      },
      "PrivacyRequestResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "request_type": {
            "title": "Request Type",
            "type": "string"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "request_type",
          "status",
          "details",
          "resolution_note",
          "due_at",
          "completed_at",
          "created_at"
        ],
        "title": "PrivacyRequestResponse",
        "type": "object"
      },
      "PrivacyRequestReview": {
        "properties": {
          "resolution_note": {
            "maxLength": 5000,
            "minLength": 3,
            "title": "Resolution Note",
            "type": "string"
          },
          "status": {
            "enum": [
              "in_review",
              "blocked_legal_hold",
              "completed",
              "denied"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "resolution_note"
        ],
        "title": "PrivacyRequestReview",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "VisaTrack API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/privacy/admin/requests": {
      "get": {
        "operationId": "list_organization_privacy_requests_api_v1_privacy_admin_requests_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrivacyRequestResponse"
                  },
                  "title": "Response List Organization Privacy Requests Api V1 Privacy Admin Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Organization Privacy Requests",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/privacy/admin/requests/{request_id}": {
      "patch": {
        "operationId": "review_privacy_request_api_v1_privacy_admin_requests__request_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacyRequestReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Privacy Request",
        "tags": [
          "privacy"
        ]
      }
    },
    "/api/v1/privacy/requests": {
      "get": {
        "operationId": "list_own_privacy_requests_api_v1_privacy_requests_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrivacyRequestResponse"
                  },
                  "title": "Response List Own Privacy Requests Api V1 Privacy Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Own Privacy Requests",
        "tags": [
          "privacy"
        ]
      },
      "post": {
        "operationId": "create_privacy_request_api_v1_privacy_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacyRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Privacy Request",
        "tags": [
          "privacy"
        ]
      }
    },
...
  }
}